[#9] Added new component file and moved code [DC]
